Correct the refcount for the system interface of the debugger
The debugger retreived and hid a pointer to the the old system interface
before replacing it with its own system interface, without updating the
reference count to indicate this additional reference.
That is now resolved as there are an addref before the replacement and a
release when finished with it.