Removing additional references to EMP
Missed some references to EMP from build mechanisms that I hadn't been
using (relied only on the CMake-based build for the tutorials).

I removed the references to EMPCore in these build files (which,
fortunately, already check for RocketCore as a replacement) and also
updated a couple of references to EMP_PLATFORM_WIN32 in the tutorials'
macros.